# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  category: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  unit: string;
  image: string;
  accent: string;
  /** @nullable */
  badge: string | null;
  featured: boolean;
}

export interface Category {
  id: string;
  name: string;
  count: number;
  icon: string;
}

export interface CatalogHighlights {
  headline: string;
  subheadline: string;
  deliveryWindow: string;
  minimumOrder: number;
  freeDeliveryThreshold: number;
}

export interface OrderItemInput {
  productId: string;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderItem {
  productId: string;
  /** @minimum 1 */
  quantity: number;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  cod: 'cod',
  upi: 'upi',
} as const;

export interface OrderInput {
  /** @minLength 16 */
  clientId: string;
  /** @minItems 1 */
  items: OrderItemInput[];
  /** @minLength 1 */
  customerName: string;
  /** @minLength 8 */
  phone: string;
  /** @minLength 5 */
  address: string;
  paymentMethod: OrderInputPaymentMethod;
}

export interface Order {
  id: string;
  clientId: string;
  items: OrderItem[];
  customerName: string;
  phone: string;
  address: string;
  status: string;
  total: number;
  eta: string;
  paymentMethod: string;
  createdAt: string;
}

export interface ProfileInput {
  /** @minLength 16 */
  clientId: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 8 */
  phone: string;
}

export interface Address {
  id: string;
  clientId: string;
  label: string;
  address: string;
  isDefault: boolean;
}

export interface AddressInput {
  /** @minLength 16 */
  clientId: string;
  /** @minLength 1 */
  label: string;
  /** @minLength 5 */
  address: string;
  isDefault?: boolean;
}

export interface AddressUpdate {
  /** @minLength 16 */
  clientId: string;
  /** @minLength 1 */
  label?: string;
  /** @minLength 5 */
  address?: string;
  isDefault?: boolean;
}

export interface Profile {
  clientId: string;
  name: string;
  phone: string;
  addresses: Address[];
}

export type ListProductsParams = {
category?: string;
search?: string;
featured?: boolean;
};

export type ListOrdersParams = {
/**
 * @minLength 16
 */
clientId: string;
};

export type GetOrderParams = {
/**
 * @minLength 1
 */
orderId: string;
/**
 * @minLength 16
 */
clientId: string;
};

export type GetProfileParams = {
/**
 * @minLength 16
 */
clientId: string;
};

export type DeleteAddressParams = {
/**
 * @minLength 1
 */
addressId: string;
/**
 * @minLength 16
 */
clientId: string;
};