# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: catalog
    description: Grocery catalog and discovery
  - name: orders
    description: Checkout and order placement
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: List grocery products
      description: Returns products with optional category and search filters.
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /categories:
    get:
      operationId: listCategories
      tags: [catalog]
      summary: List grocery categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /catalog/highlights:
    get:
      operationId: getCatalogHighlights
      tags: [catalog]
      summary: Get curated catalog highlights
      description: Returns lightweight merchandising data for the homepage.
      responses:
        "200":
          description: Curated homepage highlights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogHighlights"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a grocery order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: listOrders
      tags: [orders]
      summary: List the shopper's orders
      parameters:
        - name: clientId
          in: query
          required: true
          schema: { type: string, minLength: 16 }
      responses:
        "200":
          description: Shopper orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /order:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get one of the shopper's orders for tracking
      parameters:
        - name: orderId
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: clientId
          in: query
          required: true
          schema: { type: string, minLength: 16 }
      responses:
        "200":
          description: Order details and current tracking status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the shopper profile and saved addresses
      parameters:
        - name: clientId
          in: query
          required: true
          schema: { type: string, minLength: 16 }
      responses:
        "200":
          description: Shopper profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Create or update the shopper profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated shopper profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/addresses:
    post:
      operationId: createAddress
      tags: [profile]
      summary: Add a saved delivery address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Saved address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /profile/address:
    delete:
      operationId: deleteAddress
      tags: [profile]
      summary: Delete a saved address
      parameters:
        - name: addressId
          in: query
          required: true
          schema: { type: string, minLength: 1 }
        - name: clientId
          in: query
          required: true
          schema: { type: string, minLength: 16 }
      responses:
        "204":
          description: Address deleted
  /profile/addresses/{addressId}:
    patch:
      operationId: updateAddress
      tags: [profile]
      summary: Update or make a saved address the default
      parameters:
        - name: addressId
          in: path
          required: true
          schema: { type: string, minLength: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressUpdate"
      responses:
        "200":
          description: Updated saved address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "404":
          description: Address not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, name, category, price, unit, image, accent, badge, featured]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        unit: { type: string }
        image: { type: string }
        accent: { type: string }
        badge: { type: ["string", "null"] }
        featured: { type: boolean }
    Category:
      type: object
      required: [id, name, count, icon]
      properties:
        id: { type: string }
        name: { type: string }
        count: { type: integer }
        icon: { type: string }
    CatalogHighlights:
      type: object
      required: [headline, subheadline, deliveryWindow, minimumOrder, freeDeliveryThreshold]
      properties:
        headline: { type: string }
        subheadline: { type: string }
        deliveryWindow: { type: string }
        minimumOrder: { type: number }
        freeDeliveryThreshold: { type: number }
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1 }
    OrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1 }
    OrderInput:
      type: object
      required: [clientId, items, customerName, phone, address, paymentMethod]
      properties:
        clientId: { type: string, minLength: 16 }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
        customerName: { type: string, minLength: 1 }
        phone: { type: string, minLength: 8 }
        address: { type: string, minLength: 5 }
        paymentMethod:
          type: string
          enum: [cod, upi]
    Order:
      type: object
      required: [id, clientId, items, customerName, phone, address, status, total, eta, paymentMethod, createdAt]
      properties:
        id: { type: string }
        clientId: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        customerName: { type: string }
        phone: { type: string }
        address: { type: string }
        status: { type: string }
        total: { type: number }
        eta: { type: string }
        paymentMethod: { type: string }
        createdAt: { type: string, format: date-time }
    ProfileInput:
      type: object
      required: [clientId, name, phone]
      properties:
        clientId: { type: string, minLength: 16 }
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 8 }
    Address:
      type: object
      required: [id, clientId, label, address, isDefault]
      properties:
        id: { type: string }
        clientId: { type: string }
        label: { type: string }
        address: { type: string }
        isDefault: { type: boolean }
    AddressInput:
      type: object
      required: [clientId, label, address]
      properties:
        clientId: { type: string, minLength: 16 }
        label: { type: string, minLength: 1 }
        address: { type: string, minLength: 5 }
        isDefault: { type: boolean }
    AddressUpdate:
      type: object
      required: [clientId]
      properties:
        clientId: { type: string, minLength: 16 }
        label: { type: string, minLength: 1 }
        address: { type: string, minLength: 5 }
        isDefault: { type: boolean }
    Profile:
      type: object
      required: [clientId, name, phone, addresses]
      properties:
        clientId: { type: string }
        name: { type: string }
        phone: { type: string }
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/Address"