# lib/api-zod/src/generated/types/address.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Address {
  id: string;
  clientId: string;
  label: string;
  address: string;
  isDefault: boolean;
}