# lib/api-zod/src/generated/types/addressInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddressInput {
  /** @minLength 16 */
  clientId: string;
  /** @minLength 1 */
  label: string;
  /** @minLength 5 */
  address: string;
  isDefault?: boolean;
}